Convert ExternalLink to typescript
